Moved transfer related data types into their own file
