+++ uavcan/src/frame_assembler.rs
@@ -1,8 +1,11 @@
 use bit_field::BitField;
 
+use transfer::{
+    TransferFrame,
+};
+
 use {
     Frame,
-    TransferFrame,
     Header,
 };
